requirements.txt
abstra==3.14.2
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.11.0